move mainFragment to a fragments folder, created a new fragment for tutoraial.
