+++ app/src/main/java/com/example/techease/fragments/TutorialFragment.java
@@ -1,0 +1,25 @@
+package com.example.techease.fragments;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.techease.R;
+
+public class TutorialFragment extends Fragment {
+
+    public TutorialFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_tutorial, container, false);
+    }
+}
