+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,12 @@
     <string name="help">Help</string>
     <string name="settings">Settings</string>
     <string name="start_tutorial">Start Tutorials</string>
+
+    <string name="tutorials_instruction">Tap a button below to view tutorials in that group.</string>
+    <string name="phone_features">Phone Features</string>
+    <string name="apps_tut">Apps</string>
+    <string name="internet">Internet</string>
+    <string name="social_media">Social Media</string>
+
+
 </resources>
